format script: add MultiROM compatibility
When installing a ROM with this script inside, /system is checked and formatted depending on which filesystem is found. If we install as secondary through MultiROM system partition is still formatted because this script is executed, breaking both primary ROM and MultiROM because this last can't be used if there isn't a primary ROM inside the device. We don't want this behavior so with this update if the primary ROM is being installed/upgraded run filesystem check and format, else don't run it.

+++ prebuilt/common/bin/format.sh
@@ -13,25 +13,35 @@
 #
 # Export busybox path
 export BB=/sbin/busybox
-# Mount /system
+# Check ROM installation/upgrade:
+# if touching primary ROM run filesystem check and format on /system, else don't run it
 $BB mount /system
-# Copy mounts in a text file
-$BB mount > /tmp/mount.txt
-# Unmount /system
+export MROM=ls /tmp/META-INF/com/google/android | grep "updater-script"
 $BB umount /system
-# Set a value for FS
-export FS=$(eval $(/sbin/blkid /dev/block/mmcblk0p16 | /sbin/busybox cut -c 24-); /sbin/busybox echo $TYPE > /tmp/type.txt);
-export TYPE=`cat /tmp/type.txt | grep "ext4"`
-# Filesystem check & format
-if [ "$TYPE" == "ext4" ]; then
-	echo "Found ext4 filesystem. Formatting..."
-	mke2fs -T ext4 /dev/block/mmcblk0p16
-elif [ "$TYPE" == "" ]; then
-	echo "Found f2fs filesystem. Formatting..."
-	mkfs.f2fs /dev/block/mmcblk0p16
+if [ "$MROM" == "" ]; then
+    echo "Installing ROM on /system"
+    # Mount /system
+    $BB mount /system
+    # Copy mounts in a text file
+    $BB mount > /tmp/mount.txt
+    # Unmount /system
+    $BB umount /system
+    # Set a value for FS
+    export FS=$(eval $(/sbin/blkid /dev/block/mmcblk0p16 | /sbin/busybox cut -c 24-); /sbin/busybox echo $TYPE > /tmp/type.txt);
+    export TYPE=`cat /tmp/type.txt | grep "ext4"`
+    # Filesystem check & format
+    if [ "$TYPE" == "ext4" ]; then
+        echo "Found ext4 filesystem. Formatting..."
+        mke2fs -T ext4 /dev/block/mmcblk0p16
+    elif [ "$TYPE" == "" ]; then
+        echo "Found f2fs filesystem. Formatting..."
+        mkfs.f2fs /dev/block/mmcblk0p16
+    else
+        # Some recoveries may not have F2FS tools
+        echo "No filesystem specified. Formatting as ext4..."
+        mke2fs -T ext4 /dev/block/mmcblk0p16
+    fi
 else
-# Some recoveries may not have F2FS tools
-	echo "No filesystem specified. Formatting as ext4..."
-	mke2fs -T ext4 /dev/block/mmcblk0p16
+    echo "Installing ROM in MultiROM"
 fi
 exit 0
